Move FDC content to all claims
